Removed some legacy code
