+++ inference-engine/src/legacy_api/include/legacy/ie_util_internal.hpp
@@ -81,12 +81,4 @@ using printer_callback =
 INFERENCE_ENGINE_API_CPP(void)
 saveGraphToDot(const InferenceEngine::CNNNetwork& network, std::ostream& out, printer_callback layer_cb = nullptr);
 
-/**
-  @brief Return root data objects, i.e. objects came from input or const layers
-  @param network - network to process
-  @return set of root data objects,
-  */
-INFERENCE_ENGINE_API_CPP(std::unordered_set<DataPtr>)
-getRootDataObjects(const InferenceEngine::CNNNetwork& network);
-
 }  // namespace InferenceEngine
