+++ inference-engine/src/legacy_api/src/convert_function_to_cnn_network.cpp
@@ -54,7 +54,46 @@
 #include "legacy/graph_tools.hpp"
 #include "legacy/net_pass.h"
 #include "ie_legacy_itt.hpp"
-#include "ie_cnn_layer_builder_ngraph.h"
+
+namespace Builder {
+
+template <class T>
+std::string asString(const T& value) {
+    return std::to_string(value);
+}
+
+template <typename T>
+std::string asString(const std::vector<T>& value) {
+    std::string result;
+    for (const auto& item : value) {
+        if (!result.empty()) result += ",";
+        result += asString(item);
+    }
+    return result;
+}
+
+template <>
+std::string asString<double>(const double& value) {
+    std::ostringstream sStrm;
+    sStrm.precision(std::numeric_limits<double>::digits10);
+    sStrm << std::fixed << value;
+    std::string result = sStrm.str();
+
+    auto pos = result.find_last_not_of("0");
+    if (pos != std::string::npos) result.erase(pos + 1);
+
+    pos = result.find_last_not_of(".");
+    if (pos != std::string::npos) result.erase(pos + 1);
+
+    return result;
+}
+
+template <>
+std::string asString<float>(const float& value) {
+    return asString(static_cast<double>(value));
+}
+
+}  // namespace Builder
 
 namespace InferenceEngine {
 namespace details {
