+++ inference-engine/tests_deprecated/unit/inference_engine_tests/util_test.cpp
@@ -463,99 +463,3 @@ TEST(UtilTests, cloneNet_const) {
     ASSERT_EQ("custom_val2", getLayer(cloned, "input2")->params["custom_param2"]);
     ASSERT_EQ("custom_val3", getLayer(cloned, "input3")->params["custom_param3"]);
 }
-
-TEST(UtilTests, DISABLED_getRootDataObjects) {
-    //
-    // I1-d1-L1-d7
-    //            \
-    // I2-d2       L6
-    //      \     /
-    //       L2-d8
-    //      /
-    // I3-d3
-    //      \
-    //       L3
-    //      /
-    // C1-d4
-    //      \
-    //       L4-d9
-    //      /     \
-    // C2-d5       L7
-    //            /
-    // C3-d6-L5-d10
-    auto net = NetBuilder()
-               .data("data1",IE::TensorDesc(IE::Precision::UNSPECIFIED, IE::SizeVector{ 1,1,1 }, IE::Layout::CHW))
-               .data("data2",IE::TensorDesc(IE::Precision::UNSPECIFIED, IE::SizeVector{ 1,1,1 }, IE::Layout::CHW))
-               .data("data3",IE::TensorDesc(IE::Precision::UNSPECIFIED, IE::SizeVector{ 1,1,1 }, IE::Layout::CHW))
-               .data("data4",IE::TensorDesc(IE::Precision::UNSPECIFIED, IE::SizeVector{ 1,1,1 }, IE::Layout::CHW))
-               .data("data5",IE::TensorDesc(IE::Precision::UNSPECIFIED, IE::SizeVector{ 1,1,1 }, IE::Layout::CHW))
-               .data("data6",IE::TensorDesc(IE::Precision::UNSPECIFIED, IE::SizeVector{ 1,1,1 }, IE::Layout::CHW))
-               .data("data7",IE::TensorDesc(IE::Precision::UNSPECIFIED, IE::SizeVector{ 1,1,1 }, IE::Layout::CHW))
-               .data("data8",IE::TensorDesc(IE::Precision::UNSPECIFIED, IE::SizeVector{ 1,1,1 }, IE::Layout::CHW))
-               .data("data9",IE::TensorDesc(IE::Precision::UNSPECIFIED, IE::SizeVector{ 1,1,1 }, IE::Layout::CHW))
-               .data("data10",IE::TensorDesc(IE::Precision::UNSPECIFIED, IE::SizeVector{ 1,1,1 }, IE::Layout::CHW))
-               .layer<IE::CNNLayer>(IE::LayerParams{"input1","input",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"input2","Input",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"input3","input",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"const1","const",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"const2","Const",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"const3","const",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"layer1","dummy",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"layer2","dummy",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"layer3","dummy",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"layer4","dummy",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"layer5","dummy",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"layer6","dummy",IE::Precision::UNSPECIFIED})
-               .layer<IE::CNNLayer>(IE::LayerParams{"layer7","dummy",IE::Precision::UNSPECIFIED})
-
-               .linkToData("input1", "data1")
-               .linkToData("input2", "data2")
-               .linkToData("input3", "data3")
-
-               .linkToData("const1", "data4")
-               .linkToData("const2", "data5")
-               .linkToData("const3", "data6")
-
-               .linkDataTo("data1", "layer1")
-               .linkDataTo("data2", "layer2")
-               .linkDataTo("data3", "layer2")
-               .linkDataTo("data3", "layer3")
-
-               .linkDataTo("data4", "layer3")
-               .linkDataTo("data4", "layer4")
-               .linkDataTo("data5", "layer4")
-               .linkDataTo("data6", "layer5")
-
-               .linkToData("layer1", "data7")
-               .linkToData("layer2", "data8")
-               .linkToData("layer4", "data9")
-               .linkToData("layer5", "data10")
-
-               .linkDataTo("data7", "layer6")
-               .linkDataTo("data8", "layer6")
-
-               .linkDataTo("data9", "layer7")
-               .linkDataTo("data10", "layer7")
-
-               .addInput("data1")
-               .addInput("data2")
-               .addInput("data3")
-
-               .finalize();
-
-    auto cloned = IE::cloneNet(IE::CNNNetwork(net));
-
-    ASSERT_EQ(13, cloned->layerCount());
-    auto root_data = IE::getRootDataObjects(IE::CNNNetwork(cloned));
-    ASSERT_EQ(6, root_data.size());
-    std::unordered_set<std::string> data_names;
-    for (auto& data : root_data) {
-        data_names.insert(data->getName());
-    }
-    ASSERT_TRUE(contains(data_names, "data1"));
-    ASSERT_TRUE(contains(data_names, "data2"));
-    ASSERT_TRUE(contains(data_names, "data3"));
-    ASSERT_TRUE(contains(data_names, "data4"));
-    ASSERT_TRUE(contains(data_names, "data5"));
-    ASSERT_TRUE(contains(data_names, "data6"));
-}
